requirements.txt
neo4j==5.18.0
faker==24.0.0
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
tqdm==4.66.2
pyyaml==6.0.1
mlx-lm==0.19.2
mlx==0.20.0
huggingface-hub==0.26.2
transformers==4.47.1
peft==0.13.2
accelerate==1.1.1
scipy==1.14.1